tests/e2e-legacy/requirements.txt
# CityPulse E2E Test Suite Requirements
# Python 3.11+ required for Apache Beam compatibility

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0

# Google Cloud Libraries
google-cloud-pubsub>=2.18.0
google-cloud-bigquery>=3.11.0
google-cloud-dataflow-client>=0.8.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0

# Apache Beam (for pipeline testing) - Updated to latest stable version
apache-beam[gcp]>=2.66.0

# Data Generation and Manipulation
faker>=19.0.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
jsonschema>=4.19.0

# Logging and Monitoring
structlog>=23.1.0

# Development and Debugging
ipython>=8.14.0
ipdb>=0.13.13
